So many bugs related to the new prefix tree and new expirable chache fixed. Now seems that works fine

+++ NLPLibs/TargomanLM/libsrc/libTargomanLM/clsLMSentenceScorer.cpp
@@ -133,11 +133,11 @@ bool clsLMSentenceScorer::haveSameHistoryAs(const clsLMSentenceScorer &_oldScore
         return false;
 
     if(this->pPrivate->IndexBasedHistory.size()){
-        for(int i=0; i< MyEffectiveLenght; ++i)
+        for(size_t i=0; i< MyEffectiveLenght; ++i)
             if(this->pPrivate->IndexBasedHistory.at(i) != _oldScorer.pPrivate->IndexBasedHistory.at(i))
                 return false;
     }else{
-        for(int i=0; i< MyEffectiveLenght; ++i)
+        for(size_t i=0; i< MyEffectiveLenght; ++i)
             if(this->pPrivate->StringBasedHistory.at(i) != _oldScorer.pPrivate->StringBasedHistory.at(i))
                 return false;
     }
